chore(restructure): restructure project modules and add user management functionality
