+++ src/app.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { AppService } from './app.service';
 import { AppController } from './app.controller';
 import { GoogleAuthModule } from 'src/modules/auth/google/google-auth.module';
-import { UsersModule } from './users/users.module';
+import { UsersModule } from './modules/users/users.module';
 import { ConfigModule } from '@nestjs/config';
 import { CourseModule } from './modules/course/courses.module';
-import { DatabaseModule } from '@database/database.module';
+import { DatabaseModule } from '@common/database/database.module';
 import { AuthModule } from './modules/auth/basic/auth.module';
 import { LessonModule } from '@modules/lesson/lesson.module';
